Add f

+++ Source/Core/VideoCommon/VideoConfig_libretro.h
@@ -9639,7 +9639,7 @@ struct ini_map maps[] =
       {1,    true},                             /* ph_sznear */
       {0,    true},                             /* ph_szfar */
       {0,   false},                             /* ph_znear */
-      {(float)0.1,    true},                           /* ph_zfar */
+      {0.1f,    true},                           /* ph_zfar */
       {false, false},                           /* perf_queries */
 
       {false, false},                           /* vsync */
@@ -25703,8 +25703,8 @@ struct ini_map maps[] =
       {1,      true},                           /* projection_hack */
       {0,    true},                           /* ph_sznear */
       {0,    true},                           /* ph_szfar */
-      {(float)0.5,    true},                           /* ph_znear */
-      {(float)0.5,    true},                           /* ph_zfar */
+      {0.5f,    true},                           /* ph_znear */
+      {0.5f,    true},                           /* ph_zfar */
       {false, false},                           /* perf_queries */
 
       {false, false},                           /* vsync */
@@ -25748,8 +25748,8 @@ struct ini_map maps[] =
       {1,      true},                           /* projection_hack */
       {0,    true},                           /* ph_sznear */
       {0,    true},                           /* ph_szfar */
-      {(float)0.5,    true},                           /* ph_znear */
-      {(float)0.5,    true},                           /* ph_zfar */
+      {0.5f,    true},                           /* ph_znear */
+      {0.5f,    true},                           /* ph_zfar */
       {false, false},                           /* perf_queries */
 
       {false, false},                           /* vsync */
